fix watched market display issue with disabled pairs
